requirements.txt
faiss-cpu
gradio
numpy
pillow
timm
torch
torchvision
